virt-handler: finalize the migration only on completion
Before this patch the migration would get finalized as
soon as the domain would start running on the target, this
is wrong as finalization performs hotplug as well which can
only be done when the migration has been actually completed.

This is particularly problematic when migration runs in PostCopy
mode as the domain runs on the target while the migration is still
ongoing.

Signed-off-by: Antonio Cardace <acardace@redhat.com>